+++ azure_monitor/tests/auto_collection/live_metrics/test_manager.py
@@ -46,4 +46,36 @@ class TestLiveMetricsManager(unittest.TestCase):
                 self._manager._instrumentation_key, self._instrumentation_key
             )
             self.assertEqual(self._manager._meter, self._meter)
+            self.assertIsNotNone(self._manager._ping)
 
+    def test_switch(self):
+        """Test manager switch between ping and post."""
+        with mock.patch("requests.post"):
+            self._manager = LiveMetricsManager(
+                meter=self._meter,
+                instrumentation_key=self._instrumentation_key,
+            )
+            self._manager._ping.is_user_subscribed = True
+            self._manager.check_if_user_is_subscribed()
+            self.assertIsNone(self._manager._ping)
+            self.assertIsNotNone(self._manager._post)
+            self._manager._post.is_user_subscribed = False
+            self._manager.check_if_user_is_subscribed()
+            self.assertIsNone(self._manager._post)
+            self.assertIsNotNone(self._manager._ping)
+
+    def test_ping(self):
+        """Test ping send requests to Live Metrics service."""
+        with mock.patch("requests.post") as request:
+            self._manager = LiveMetricsManager(
+                meter=self._meter,
+                instrumentation_key=self._instrumentation_key,
+            )
+            self._manager._ping.is_user_subscribed = True
+            self._manager.check_if_user_is_subscribed()
+            self.assertIsNone(self._manager._ping)
+            self.assertIsNotNone(self._manager._post)
+            self._manager._post.is_user_subscribed = False
+            self._manager.check_if_user_is_subscribed()
+            self.assertIsNone(self._manager._post)
+            self.assertIsNotNone(self._manager._ping)
